Move some logic to builder
