+++ python-frontend/src/main/java/org/sonar/python/semantic/v2/FunctionTypeBuilder.java
@@ -1,0 +1,159 @@
+/*
+ * SonarQube Python Plugin
+ * Copyright (C) 2011-2024 SonarSource SA
+ * mailto:info AT sonarsource DOT com
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+package org.sonar.python.semantic.v2;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
+import javax.annotation.Nullable;
+import org.sonar.plugins.python.api.tree.AnyParameter;
+import org.sonar.plugins.python.api.tree.FunctionDef;
+import org.sonar.plugins.python.api.tree.Name;
+import org.sonar.plugins.python.api.tree.ParameterList;
+import org.sonar.plugins.python.api.tree.Token;
+import org.sonar.plugins.python.api.tree.Tree;
+import org.sonar.python.tree.TreeUtils;
+import org.sonar.python.types.v2.FunctionType;
+import org.sonar.python.types.v2.ParameterV2;
+import org.sonar.python.types.v2.PythonType;
+
+import static org.sonar.python.tree.TreeUtils.locationInFile;
+
+public class FunctionTypeBuilder {
+
+  private boolean hasVariadicParameter;
+  private String name;
+  private List<PythonType> attributes;
+  private List<ParameterV2> parameters;
+  private boolean isAsynchronous;
+  private boolean hasDecorators;
+  private boolean isInstanceMethod;
+  private PythonType owner;
+  private PythonType returnType = PythonType.UNKNOWN;
+
+  private static final String CLASS_METHOD_DECORATOR = "classmethod";
+  private static final String STATIC_METHOD_DECORATOR = "staticmethod";
+
+  public FunctionTypeBuilder fromFunctionDef(FunctionDef functionDef) {
+    this.name = functionDef.name().name();
+    this.attributes = new ArrayList<>();
+    this.parameters = new ArrayList<>();
+    isAsynchronous = functionDef.asyncKeyword() != null;
+    hasDecorators = !functionDef.decorators().isEmpty();
+    isInstanceMethod = isInstanceMethod(functionDef);
+    ParameterList parameterList = functionDef.parameters();
+    if (parameterList != null) {
+      createParameterNames(parameterList.all(), null);
+    }
+    return this;
+  }
+
+  public FunctionType build() {
+    return new FunctionType(name, attributes, parameters, returnType, isAsynchronous, hasDecorators, isInstanceMethod, hasVariadicParameter, owner);
+  }
+
+  private static boolean isInstanceMethod(FunctionDef functionDef) {
+    return !functionDef.name().name().equals("__new__") && functionDef.isMethodDefinition() && functionDef.decorators().stream()
+      .map(decorator -> TreeUtils.decoratorNameFromExpression(decorator.expression()))
+      .filter(Objects::nonNull)
+      .noneMatch(decorator -> decorator.equals(STATIC_METHOD_DECORATOR) || decorator.equals(CLASS_METHOD_DECORATOR));
+  }
+
+  public void setOwner(PythonType owner) {
+    this.owner = owner;
+  }
+
+  private void createParameterNames(List<AnyParameter> parameterTrees, @Nullable String fileId) {
+    FunctionType.ParameterState parameterState = new FunctionType.ParameterState();
+    parameterState.positionalOnly = parameterTrees.stream().anyMatch(param -> Optional.of(param)
+      .filter(p -> p.is(Tree.Kind.PARAMETER))
+      .map(p -> ((org.sonar.plugins.python.api.tree.Parameter) p).starToken())
+      .map(Token::value)
+      .filter("/"::equals)
+      .isPresent()
+    );
+    for (AnyParameter anyParameter : parameterTrees) {
+      if (anyParameter.is(Tree.Kind.PARAMETER)) {
+        addParameter((org.sonar.plugins.python.api.tree.Parameter) anyParameter, fileId, parameterState);
+      } else {
+        parameters.add(new ParameterV2(null, PythonType.UNKNOWN, null, false, parameterState, false, false, null, locationInFile(anyParameter, fileId)));
+      }
+    }
+  }
+
+  private void addParameter(org.sonar.plugins.python.api.tree.Parameter parameter, @Nullable String fileId, FunctionType.ParameterState parameterState) {
+    Name parameterName = parameter.name();
+    Token starToken = parameter.starToken();
+    if (parameterName != null) {
+      FunctionType.ParameterType parameterType = getParameterType(parameter);
+      this.parameters.add(new ParameterV2(parameterName.name(), parameterType.pythonType(), "UNKNOWN", parameter.defaultValue() != null,
+        parameterState, parameterType.isKeywordVariadic(), parameterType.isPositionalVariadic(), null, locationInFile(parameter, fileId)));
+      if (starToken != null) {
+        hasVariadicParameter = true;
+        parameterState.keywordOnly = true;
+        parameterState.positionalOnly = false;
+      }
+    } else if (starToken != null) {
+      if ("*".equals(starToken.value())) {
+        parameterState.keywordOnly = true;
+        parameterState.positionalOnly = false;
+      }
+      if ("/".equals(starToken.value())) {
+        parameterState.positionalOnly = false;
+      }
+    }
+  }
+
+  private FunctionType.ParameterType getParameterType(org.sonar.plugins.python.api.tree.Parameter parameter) {
+    // TODO: SONARPY-1773 handle parameter declared types
+    PythonType pythonType = PythonType.UNKNOWN;
+    boolean isPositionalVariadic = false;
+    boolean isKeywordVariadic = false;
+    Token starToken = parameter.starToken();
+    if (starToken != null) {
+      // https://docs.python.org/3/reference/compound_stmts.html#function-definitions
+      hasVariadicParameter = true;
+      if ("*".equals(starToken.value())) {
+        // if the form “*identifier” is present, it is initialized to a tuple receiving any excess positional parameters
+        isPositionalVariadic = true;
+        // Should set type to TUPLE
+        pythonType = PythonType.UNKNOWN;
+      }
+      if ("**".equals(starToken.value())) {
+        //  If the form “**identifier” is present, it is initialized to a new ordered mapping receiving any excess keyword arguments
+        isKeywordVariadic = true;
+        // Should set type to DICT
+        pythonType = PythonType.UNKNOWN;
+      }
+    } else {
+      // Should resolve type annotation
+      pythonType = PythonType.UNKNOWN;
+    }
+    return new FunctionType.ParameterType(pythonType, isKeywordVariadic, isPositionalVariadic);
+  }
+
+  static class ParameterState {
+    boolean keywordOnly = false;
+    boolean positionalOnly = false;
+  }
+
+  record ParameterType(PythonType pythonType, boolean isKeywordVariadic, boolean isPositionalVariadic) { }
+}
